Python_Development_Tasks/L2_Intermediate/Task_2_Data_Scraper/requirements.txt

requests==2.32.5
python-dotenv==1.2.2
beautifulsoup4==4.14.3
fake-useragent==2.2.0
